validation: moving CCSP/validated OCP parts from OPCT docs
This change introduce the /validation path with steps to
partners validated the OpenShift installations in
non-integrated providers.

Those steps are used by CCSP/Validated[1] program, which use
the tool OPCT[2] to run conformance tests.

Those guides was initialy created in the OPCT tool[3], but it
must be moved to a single entry point to partners, leaving
the tool to its documentation.

[1] https://source.redhat.com/groups/public/global_partner_engagement/global_partner_engagement_wiki/validated_ccsp
[2] https://redhat-openshift-ecosystem.github.io/provider-certification-tool/
[3] https://redhat-openshift-ecosystem.github.io/provider-certification-tool/user/